+++ composer.json
@@ -1,6 +1,6 @@
 {
     "name": "initphp/http",
-    "description": "InitPHP PSR-7 HTTP Message Library",
+    "description": "HTTP",
     "type": "library",
     "license": "MIT",
     "autoload": {
@@ -19,7 +19,9 @@
     "minimum-stability": "stable",
     "require": {
         "php": ">=7.4",
+        "ext-json": "*",
         "psr/http-message": "^1.0",
-        "psr/http-factory": "^1.0"
+        "psr/http-factory": "^1.0",
+        "psr/http-client": "^1.0"
     }
 }
