+++ src/Emitter/Exceptions/EmitHeaderException.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * EmitHeaderException.php
+ *
+ * This file is part of InitPHP HTTP.
+ *
+ * @author     Muhammet ŞAFAK <info@muhammetsafak.com.tr>
+ * @copyright  Copyright © 2022 Muhammet ŞAFAK
+ * @license    ./LICENSE  MIT
+ * @version    2.0
+ * @link       https://www.muhammetsafak.com.tr
+ */
+
+declare(strict_types=1);
+
+namespace InitPHP\HTTP\Emitter\Exceptions;
+
+class EmitHeaderException extends \RuntimeException
+{
+}
