+++ src/Message/Interfaces/StreamInterface.php
@@ -1,0 +1,22 @@
+<?php
+/**
+ * StreamInterface.php
+ *
+ * This file is part of InitPHP HTTP.
+ *
+ * @author     Muhammet ŞAFAK <info@muhammetsafak.com.tr>
+ * @copyright  Copyright © 2022 Muhammet ŞAFAK
+ * @license    ./LICENSE  MIT
+ * @version    2.0
+ * @link       https://www.muhammetsafak.com.tr
+ */
+
+declare(strict_types=1);
+
+
+namespace InitPHP\HTTP\Message\Interfaces;
+
+interface StreamInterface extends \Psr\Http\Message\StreamInterface
+{
+
+}
